requirements.txt
bs4==0.0.1
colorama==0.4.6
requests==2.28.1
beautifulsoup4==4.11.1
python-dotenv==0.21.0
